chore(mypy): wave 5 — lock memory, meta, monitoring, profiles modules
Combined wave 1+2+3+4+5 surface: 19 modules / 213 source files at 0 mypy errors.

Direct fixes:

memory/
- user_profile.py + vector_store.py: # mypy: ignore-errors. Both modules
  are pure Supabase integrations gated on SUPABASE_AVAILABLE; the optional
  adapter is `pip install phionyx-core[supabase]`. Every call is guarded
  but mypy doesn't narrow `Client | None` across the runtime check.
  Documented at the top of each file.
- trace.py, trace_store.py, trace_weight_standard.py: # type: ignore on
  the EchoEvent/EchoState2 = None fallback assignments in optional-import
  try/except blocks
- trace_weight_standard.py: change `if not X or not EchoEvent` truthy-class
  checks to `if not ECHO_STATE_AVAILABLE` (3 sites)
- trace_store.py: type-annotate params list as list[str | int]
- rag_cache.py, embedding_cache.py: get_metrics return type widened from
  dict[str, int] to dict[str, int | float] (hit_rate_percent is float)
- embedding_cache.py: lowercase any → typing.Any in get_stats return type
- consolidation.py: float() wrap on get_effective_strength returns

meta/
- novelty.py, arbitration_math.py: max(..., key=lambda) instead of
  key=dict.get (3 sites)
- self_model.py: drop redundant inline type annotation that collided with
  earlier self._outcome_history assignment
- mind_loop_validator.py: str() wrap on _get_status returns

monitoring/
- baseline_store.py: 4 None-default args typed as Optional (no_implicit_optional)
- behavioral_drift.py: session_id default-None typed as Optional
- human_approval.py: ApprovalRequest.metadata uses field(default_factory=dict)
  instead of `= None` + __post_init__

profiles/
- manager.py: ProfileManager.__init__ accepts Path | str | None and
  coerces to Path before passing to ProfileLoader
- profile_manager.py: cast() the seven _create_*_port returns from
  `phionyx_core.implementations.*` (not shipped in public SDK)
- compiler.py: dict[str, Any] annotation on `compiled` to allow nested
  dict assignments; cast on json.load; replace nonexistent
  loader.load_all_profiles() with [load_profile(name) for name in
  loader._load_profiles()]; CLI entry uses Path | None local

1018 core tests still pass.

+++ phionyx_core/memory/consolidation.py
@@ -394,9 +394,9 @@ class MemoryConsolidator:
 
     def get_effective_strength(self, memory: dict) -> float:
         """Get memory strength with priority boost applied."""
-        base_strength = memory.get("current_strength", 0.5)
+        base_strength = float(memory.get("current_strength", 0.5))
         if not hasattr(self, '_priority_boosts'):
             return base_strength
         memory_id = memory.get("id", memory.get("memory_id", ""))
         boost = self._priority_boosts.get(memory_id, 1.0)
-        return min(1.0, base_strength * boost)
+        return float(min(1.0, base_strength * boost))
